requirements.txt
PyQt5==5.15.7
pyserial==3.5